+++ docs/source/change_log.rst
@@ -4,6 +4,7 @@ Change Log
 ``Master branch``
 ^^^^^^^^^^^^^^^^^
 * Fix the import issue, close `#76 <https://github.com/gicait/geoserver-rest/issues/76>`_
+* Removed the ``rest`` from Geoserver class URL, Revert back to previous state. Close `#77 <https://github.com/gicait/geoserver-rest/issues/76>`_
 
 
 ``[v2.3.0 - 2022-05-06]``